add validation

+++ plugin/vim-rsync.vim
@@ -9,21 +9,49 @@ function! RsyncExclude(project)
     return ""
 endfunction
 
+function!RsyncValidate(project)
+  if !has_key(a:project, "user")
+    echoerr "The field 'user' in g:vim_rsync cannot be empty"
+    return 0
+  endif
+
+  if !has_key(a:project, "ip_address")
+    echoerr "The field 'ip_address' in g:vim_rsync cannot be empty"
+    return 0
+  endif
+
+  if !has_key(a:project, "local_directory")
+    echoerr "The field 'local_directory' in g:vim_rsync cannot be empty"
+    return 0
+  endif
+
+  if !has_key(a:project, "remote_directory")
+    echoerr "The field 'remote_directory' in g:vim_rsync cannot be empty"
+    return 0
+  endif
+
+  return 1
+endfunction
+
 function! RsyncPush(args)
   if has_key(g:vim_rsync, a:args)
     let project = g:vim_rsync[a:args]
-    exe "!rsync -r --delete " . RsyncExclude(project) . " " . project["local_directory"] . " " . project["user"] . "@" . project["ip_address"] . ":" . project["remote_directory"]
+    if RsyncValidate(project)
+      exe "!rsync -r --delete " . RsyncExclude(project) . " " . project["local_directory"] . " " . project["user"] . "@" . project["ip_address"] . ":" . project["remote_directory"]
+    endif
   else
-    echo "The project is not defined in vim_rsync"
+    echoerr "The project '" . a:args . "' is not defined in g:vim_rsync"
   endif
 endfunction
 
 function! RsyncPull(args)
   if has_key(g:vim_rsync, a:args)
     let project = g:vim_rsync[a:args]
-    exe "!rsync -r --delete " . RsyncExclude(project) . " " . project["user"] . "@" . project["ip_address"] . ":" . project["remote_directory"] . " " . project["local_directory"]
+    if RsyncValidate(project)
+      exe "!rsync -r --delete " . RsyncExclude(project) . " " . project["user"] . "@" . project["ip_address"] . ":" . project["remote_directory"] . " " . project["local_directory"]
+    endif
   else
-    echo "The project is not defined in vim_rsync"
+    echoerr "The project '" . a:args . "' is not defined in g:vim_rsync"
   endif
 endfunction
